+++ CHANGELOG.rst
@@ -6,6 +6,7 @@ Changelog
 
 * :issue:`2`: Fix Read the Docs automodule not working properly.
 * :issue:`3`: Add ``sphinx-issues`` extension to Read the Docs.
+* Add the ``VcfFrame.combine`` and ``VcfFrame.subtract`` methods.
 
 0.1.4 (2021-04-21)
 ------------------